@@ -19,83 +18,61 @@
     }
   },
   "dependencies": {
-    "@types/classnames": "2.2.7",
-    "@types/lodash.findindex": "4.6.6",
-    "@types/lodash.isnil": "4.0.6",
-    "animate.css": "3.7.0",
-    "axios": "0.18.0",
-    "bootstrap": "4.3.1",
-    "classnames": "^2.2.6",
-    "lodash.findindex": "4.6.0",
-    "lodash.isnil": "4.0.0",
-    "vue": "2.6.10",
-    "vue-i18n": "8.9.0",
-    "vue-router": "3.0.2",
-    "vuex": "^3.1.0"
+    "@types/classnames": "2.3.1",
+    "animate.css": "4.1.1",
+    "axios": "0.27.2",
+    "bootstrap": "5.1.3",
+    "classnames": "2.3.1",
+    "vue": "3.2.37",
+    "vue-i18n": "9.1.10",
+    "vue-router": "4.0.16",
+    "vuex": "4.0.2",
+    "vuex-composition-helpers": "^2.0.3"
   },
   "devDependencies": {
-    "@babel/core": "7.4.3",
-    "@babel/plugin-proposal-class-properties": "7.4.0",
-    "@babel/plugin-proposal-object-rest-spread": "7.4.3",
-    "@babel/plugin-syntax-async-generators": "7.2.0",
-    "@babel/plugin-syntax-dynamic-import": "7.2.0",
-    "@babel/plugin-transform-async-to-generator": "7.4.0",
-    "@babel/plugin-transform-modules-commonjs": "7.4.3",
-    "@babel/plugin-transform-runtime": "7.4.3",
-    "@babel/preset-env": "7.4.3",
-    "@babel/preset-flow": "7.0.0",
-    "@babel/preset-react": "7.0.0",
-    "@babel/register": "7.4.0",
-    "@babel/runtime": "7.4.3",
-    "@typescript-eslint/eslint-plugin": "1.6.0",
-    "@typescript-eslint/parser": "1.6.0",
-    "add-asset-html-webpack-plugin": "3.1.3",
-    "autoprefixer": "9.5.0",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-loader": "8.0.5",
-    "commitizen": "3.0.7",
-    "compression-webpack-plugin": "2.0.0",
-    "copy-webpack-plugin": "5.0.2",
-    "css-loader": "2.1.1",
-    "cz-conventional-changelog": "2.1.0",
-    "escape-string-regexp": "1.0.5",
-    "eslint": "5.16.0",
-    "eslint-config-airbnb-base": "13.1.0",
-    "eslint-import-resolver-webpack": "0.11.0",
-    "eslint-loader": "2.1.2",
-    "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-vue": "5.2.2",
-    "eslint-plugin-vue-a11y": "0.0.28",
-    "file-loader": "3.0.1",
-    "hard-source-webpack-plugin": "0.13.1",
-    "html-webpack-plugin": "4.0.0-beta.4",
-    "interpolate-html-plugin": "3.0.0",
-    "json-server": "^0.14.2",
-    "mini-css-extract-plugin": "0.5.0",
-    "node-sass": "4.11.0",
+    "@typescript-eslint/eslint-plugin": "5.29.0",
+    "@typescript-eslint/parser": "5.29.0",
+    "@vue/compiler-sfc": "3.2.37",
+    "add-asset-html-webpack-plugin": "5.0.2",
+    "commitizen": "^4.2.4",
+    "compression-webpack-plugin": "10.0.0",
+    "css-loader": "6.7.1",
+    "css-minimizer-webpack-plugin": "4.0.0",
+    "cz-conventional-changelog": "^3.3.0",
+    "eslint": "8.18.0",
+    "eslint-config-airbnb-base": "15.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "2.7.1",
+    "eslint-import-resolver-webpack": "0.13.2",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-vue": "9.1.1",
+    "eslint-plugin-vue-a11y": "0.0.31",
+    "eslint-webpack-plugin": "3.1.1",
+    "file-loader": "6.2.0",
+    "html-webpack-plugin": "5.5.0",
+    "json-server": "0.17.0",
+    "mini-css-extract-plugin": "2.6.1",
+    "node-sass": "7.0.1",
     "node-uuid": "1.4.8",
-    "optimize-css-assets-webpack-plugin": "5.0.1",
-    "postcss-loader": "3.0.0",
-    "postcss-preset-env": "6.6.0",
+    "postcss-loader": "7.0.0",
     "precss": "4.0.0",
-    "rimraf": "2.6.3",
+    "prettier": "^2.7.1",
+    "rimraf": "3.0.2",
     "run-sequence": "2.2.1",
-    "sass-loader": "7.1.0",
-    "script-ext-html-webpack-plugin": "2.1.3",
-    "style-loader": "0.23.1",
-    "ts-loader": "5.3.3",
-    "typescript": "3.4.2",
-    "url-loader": "1.1.2",
-    "vue-eslint-parser": "6.0.3",
-    "vue-loader": "15.7.0",
-    "vue-style-loader": "4.1.2",
-    "vue-template-compiler": "2.6.10",
-    "webpack": "4.29.6",
-    "webpack-bundle-analyzer": "3.2.0",
-    "webpack-cli": "3.3.0",
-    "webpack-dev-server": "3.2.1",
-    "webpack-dll-bundles-plugin": "1.0.0-beta.5",
-    "webpack-merge": "4.2.1",
-    "yargs": "13.2.2"
+    "sass-loader": "13.0.0",
+    "script-ext-html-webpack-plugin": "2.1.5",
+    "style-loader": "3.3.1",
+    "ts-loader": "9.3.0",
+    "typescript": "4.7.4",
+    "url-loader": "4.1.1",
+    "vue-eslint-parser": "9.0.2",
+    "vue-loader": "next",
+    "vue-style-loader": "4.1.3",
+    "webpack": "5.73.0",
+    "webpack-bundle-analyzer": "4.5.0",
+    "webpack-cli": "4.10.0",
+    "webpack-dev-server": "4.9.2",
+    "webpack-merge": "5.8.0",
+    "yargs": "17.5.1"
   }
 }
