@@ -4,25 +4,23 @@
       <button
         v-for="filter in filters"
         :key="filter.id"
-        @click="handleFilterClick(filter)"
         class="toolbar__button"
+        @click="handleFilterClick(filter)"
       >
-        <Icon
-          :name="filter.icon"
-          :selected="filter.id === selected"
-        />
+        <Icon :name="filter.icon" :selected="filter.id === selected" />
       </button>
     </div>
   </nav>
 </template>
 
 <script lang="ts">
-import Vue from 'vue';
+import { defineComponent } from 'vue';
 import Icon from '@/components/base/Icon';
 
 import { IDataToolbar, Filter } from './toolbar.types';
 
-export default Vue.extend({
+export default defineComponent({
+  name: 'ToolBar',
   components: {
     Icon,
   },
@@ -53,28 +51,27 @@ export default Vue.extend({
 </script>
 
 <style lang="scss">
-  @import '../../assets/scss/mixins/flex';
+@import '../../assets/scss/mixins/flex';
 
-  .toolbar {
-    @include flex($justify-content: flex-end);
-    padding: 0.5rem 0;
+.toolbar {
+  @include flex($justify-content: flex-end);
+  padding: 0.5rem 0;
+}
+
+.toolbar__button {
+  background: none;
+  margin: 0;
+  border: none;
+  outline: none;
+  padding: 0;
+  &:focus {
+    border: none !important;
+    outline: none !important;
   }
+}
 
-  .toolbar__button {
-    background: none;
-    margin: 0;
-    border: none;
-    outline: none;
-    padding: 0;
-    &:focus {
-      border: none !important;
-      outline: none !important;
-    }
-  }
-
-  .toolbar__filters {
-    @include flex($justify-content: space-evenly);
-    width: 5rem;
-  }
-
+.toolbar__filters {
+  @include flex($justify-content: space-evenly);
+  width: 5rem;
+}
 </style>
