@@ -1,13 +1,6 @@
 <template>
-  <div
-    role="listbox"
-    class="select"
-  >
-    <div
-      @click="handleClick"
-      role="button"
-      class="custom-select"
-    >
+  <div role="listbox" class="select">
+    <div @click="handleClick" role="button" class="custom-select">
       {{ buttonText }}
     </div>
     <div v-if="isOpen">
@@ -18,11 +11,7 @@
         aria-labelledby="selectBox"
         class="select__items"
       >
-        <div
-          @click="handleSelect"
-          role="option"
-          class="select__item"
-        >
+        <div @click="handleSelect" role="option" class="select__item">
           {{ item.text }}
         </div>
       </div>
@@ -35,7 +24,7 @@ import Vue from 'vue';
 
 import { ISelectItem } from './select.types';
 
-export default Vue.extend({
+export default defineComponent({
   name: 'Select',
   props: {
     items: {
@@ -66,14 +55,14 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scope>
-  @import '~styles/functions/px-to-rem';
-  @import '~styles/mixins';
+@import '~styles/functions/px-to-rem';
+@import '~styles/mixins';
 
-  .select {
-    @include element(item) {
-      background: var(--white);
-      padding: 6px 14px;
-      border-bottom: 1px solid #ddd;
-    }
+.select {
+  @include element(item) {
+    background: var(--white);
+    padding: 6px 14px;
+    border-bottom: 1px solid #ddd;
   }
+}
 </style>
