@@ -3,11 +3,7 @@
     <div class="row">
       <div class="col-12 col-sm-12 col-md-12 col-md-9 col-xl-9">
         <figure class="product-image">
-          <img
-            :src="product.img"
-            :alt="product.name"
-            class="img-fluid"
-          >
+          <img :src="product.img" :alt="product.name" class="img-fluid" />
         </figure>
       </div>
       <div class="col-12 col-sm-12 col-md-12 col-md-3 col-xl-3">
@@ -26,7 +22,7 @@ import ProductDescription from './ProductDescription';
 
 const { mapState, mapActions } = createNamespacedHelpers('products');
 
-export default Vue.extend({
+export default defineComponent({
   name: 'Product',
   components: {
     ProductDescription,
@@ -41,18 +37,18 @@ export default Vue.extend({
 </script>
 
 <style lang="scss">
-  @import '~styles/variables';
-  @import '~styles/mixins';
+@import '~styles/variables';
+@import '~styles/mixins';
 
-  .product {
-    padding-top: 4rem;
+.product {
+  padding-top: 4rem;
+}
+
+.product-image {
+  @include flex($justify-content: center);
+
+  @include media($lg) {
+    justify-content: flex-start;
   }
-
-  .product-image {
-    @include flex($justify-content: center);
-
-    @include media($lg) {
-      justify-content: flex-start;
-    }
-  }
+}
 </style>
