@@ -18,12 +18,12 @@ import Vue from 'vue';
 
 import { createNamespacedHelpers } from 'vuex';
 
-import ProductCard from '@/components/ProductCard.vue';
-import Toolbar from '@/components/Toolbar';
+import ProductCard from 'components/ProductCard.vue';
+import Toolbar from 'components/Toolbar';
 
 const { mapState, mapActions } = createNamespacedHelpers('products');
 
-export default Vue.extend({
+export default defineComponent({
   components: {
     ProductCard,
     Toolbar,
@@ -47,8 +47,7 @@ export default Vue.extend({
 </script>
 
 <style lang="scss">
-
-  .shop {
-    position: relative;
-  }
+.shop {
+  position: relative;
+}
 </style>
