@@ -1,10 +1,10 @@
-const webpackMerge = require('webpack-merge');
+const { merge: webpackMerge } = require('webpack-merge');
 
 const loadPresets = (env = { presets: [] }) => {
   const presets = env.presets || [];
 
   const mergedPresets = [].concat(...[presets]);
-  const mergedConfigs = mergedPresets.map(presetName => {
+  const mergedConfigs = mergedPresets.map((presetName) => {
     return require(`./presets/webpack.${presetName}`)(env);
   });
 
