@@ -8,11 +8,9 @@ const resolveApp = (relativePath) => path.resolve(appDirectory, relativePath);
 const paths = {
   src: resolveApp('src'),
   dist: resolveApp('dist'),
+  indexHTML: resolveApp('src/index.html'),
 };
 
-paths.resolveModules = [
-  paths.src,
-  'node_modules',
-];
+paths.resolveModules = [paths.src, 'node_modules'];
 
 module.exports = paths;
