@@ -1,62 +1,35 @@
-const webpack = require('webpack');
+const ESLintPlugin = require('eslint-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const VueLoaderPlugin = require('vue-loader/lib/plugin');
-const postcssPresetEnv = require('postcss-preset-env');
-
-const { setupPath } = require('./helpers');
+const { VueLoaderPlugin } = require('vue-loader');
+const webpack = require('webpack');
 const paths = require('./paths');
-
-const progressHandler = (percentage, message, ...args) => {
-  // eslint-disable-next-line no-console
-  console.info(percentage, message, ...args);
-};
 
 module.exports = (mode) => {
   const prodMode = mode === 'production';
-  return {
 
+  return {
     entry: ['./src/main.ts'],
 
     resolve: {
       alias: {
-        vue$: 'vue/dist/vue.esm.js',
+        /**
+         * @see https://github.com/vuejs/core/tree/main/packages/vue#with-a-bundler
+         */
+        vue$: 'vue/dist/vue.esm-bundler.js',
         api$: `${paths.src}/api/api.ts`,
         '@': `${paths.src}`,
-        assets: `${paths.src}/assets`,
         styles: `${paths.src}/assets/scss`,
-        constants: `${paths.src}/constants`,
-        core: `${paths.src}/components/core`,
-        '@core': `${paths.src}/components/core`,
-        '@store': `${paths.src}/store`,
-        utils: `${paths.src}/utils`,
       },
       extensions: ['.js', '.vue', '.json', '.ts'],
     },
 
     module: {
-      // rules for modules (configure loaders, parser options, etc.)
       rules: [
-        {
-          enforce: 'pre',
-          test: /\.(js|vue)$/,
-          loader: 'eslint-loader',
-          exclude: /node_modules/,
-        },
         {
           test: /\.vue$/,
           loader: 'vue-loader',
         },
-        // {
-        //   test: /\.js?$/,
-        //   exclude: [/node_modules/],
-        //   use: [
-        //     {
-        //       loader: 'babel-loader',
-        //     },
-        //   ],
-        // },
         {
           test: /\.ts$/,
           loader: 'ts-loader',
@@ -75,16 +48,10 @@ module.exports = (mode) => {
               loader: 'css-loader',
             },
             {
-              loader: 'postcss-loader', // Run post css actions
+              loader: 'postcss-loader',
               options: {
-                plugins() { // post css plugins, can be exported to postcss.config.js
-                  return [
-                    // eslint-disable-next-line global-require
-                    require('precss'),
-                    // eslint-disable-next-line global-require
-                    require('autoprefixer'),
-                    postcssPresetEnv(),
-                  ];
+                postcssOptions: {
+                  plugins: [require('precss')],
                 },
               },
             },
@@ -103,38 +70,44 @@ module.exports = (mode) => {
         },
         {
           test: /.(ttf|otf|eot|svg|woff(2)?)(\?[a-z0-9]+)?$/,
-          use: [{
-            loader: 'file-loader',
-            options: {
-              name: '[name].[ext]',
-              outputPath: 'fonts/', // where the fonts will go
+          use: [
+            {
+              loader: 'file-loader',
+              options: {
+                name: '[name].[ext]',
+                outputPath: 'fonts/', // where the fonts will go
+              },
             },
-          }],
+          ],
         },
       ],
     },
     plugins: [
       new VueLoaderPlugin(),
-      new webpack.ProgressPlugin(progressHandler),
+
       new MiniCssExtractPlugin({
         // Options similar to the same options in webpackOptions.output
         // both options are optional
         filename: prodMode ? '[name].[hash].css' : '[name].css',
         chunkFilename: prodMode ? '[id].[hash].css' : '[id].css',
       }),
+
       new HtmlWebpackPlugin({
-        template: setupPath('../src/index.html'),
+        template: paths.indexHTML,
       }),
-      // copy files and folders to specific paths.
-      new CopyWebpackPlugin([{
-        // Copy `assets` contents to {output}/assets/
-        from: 'src/assets',
-        to: 'assets',
-        ignore: [
-          // Doesn't copy any files with a scss extension
-          '*.scss',
-        ],
-      }]),
+
+      /**
+       * @see https://github.com/vuejs/core/tree/main/packages/vue#bundler-build-feature-flags
+       */
+      new webpack.DefinePlugin({
+        __VUE_OPTIONS_API__: false,
+        __VUE_PROD_DEVTOOLS__: false,
+      }),
+
+      new ESLintPlugin({
+        extensions: ['vue', 'ts'],
+        failOnError: false,
+      }),
     ],
   };
 };
