@@ -1,11 +1,10 @@
 const Webpack = require('webpack');
-const webpackMerge = require('webpack-merge');
+const { merge: webpackMerge } = require('webpack-merge');
 const ScriptExtHtmlWebpackPlugin = require('script-ext-html-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const CssMinimizerPlugin = require('css-minimizer-webpack-plugin');
 const CompressionPlugin = require('compression-webpack-plugin');
 
-const { setupPath } = require('./helpers');
 const paths = require('./paths');
 
 module.exports = ({ mode }) => {
@@ -27,29 +26,29 @@ module.exports = ({ mode }) => {
       hints: 'warning', // enum
       maxAssetSize: 200000, // int (in bytes),
       maxEntrypointSize: 400000, // int (in bytes)
-      assetFilter: assetFilename => (
+      assetFilter: (assetFilename) =>
         // Function predicate that provides asset filenames
-        assetFilename.endsWith('.css') || assetFilename.endsWith('.js')
-      ),
+        assetFilename.endsWith('.css') || assetFilename.endsWith('.js'),
+    },
+
+    optimization: {
+      minimizer: [
+        new CssMinimizerPlugin({
+          minimizerOptions: {
+            preset: [
+              'default',
+              {
+                discardComments: { removeAll: true },
+              },
+            ],
+          },
+        }),
+      ],
     },
 
     plugins: [
-      new OptimizeCssAssetsPlugin({
-        assetNameRegExp: /\.css$/g,
-        // eslint-disable-next-line
-        cssProcessor: require('cssnano'),
-        cssProcessorPluginOptions: {
-          preset: [
-            'default', {
-              discardComments: {
-                removeAll: true,
-              },
-            }],
-        },
-        canPrint: true,
-      }),
       new HtmlWebpackPlugin({
-        template: setupPath('./templates/index.html'),
+        template: paths.indexHTML,
         minify: {
           collapseWhitespace: true,
           keepClosingSlash: true,
