@@ -1,19 +1,28 @@
-<template functional>
+<template>
   <img
-    :src="`assets/svg/icons/${props.name}${props.selected ? '-selected' : ''}.svg`"
+    :src="`assets/svg/icons/${name}${selected ? '-selected' : ''}.svg`"
     :style="{
-      height: props.height ? props.height : '1.5rem',
-      width: props.width ? props.width : '1.5rem',
+      height: height ? height : '1.5rem',
+      width: width ? width : '1.5rem',
     }"
-    :alt="props.name"
+    :alt="name"
     class="icon"
-  >
+  />
 </template>
 
+<script setup lang="ts">
+defineProps<{
+  name?: string;
+  selected?: boolean;
+  height?: string;
+  width?: string;
+}>();
+</script>
+
 <style lang="scss">
-  .icon {
-    cursor: pointer;
-    height: 1.5rem;
-    width: 1.5rem;
-  }
+.icon {
+  cursor: pointer;
+  height: 1.5rem;
+  width: 1.5rem;
+}
 </style>
