@@ -4,13 +4,11 @@
     :value="value"
     @input="$emit('input', $event.target.value)"
     class="input"
-  >
+  />
 </template>
 
-<script lang="ts">
-import Vue from 'vue';
-
-export default Vue.extend({
+<script>
+export default {
   props: {
     value: {
       type: [String, Number],
@@ -21,18 +19,18 @@ export default Vue.extend({
       default: '',
     },
   },
-});
+};
 </script>
 
 <style lang="scss">
-  .input {
-    appearance: none;
-    background: transparent;
-    border-radius: 2px;
-    border: 2px solid #ddd;
-    position: relative;
-    &:focus {
-      outline:none;
-    }
+.input {
+  appearance: none;
+  background: transparent;
+  border-radius: 2px;
+  border: 2px solid #ddd;
+  position: relative;
+  &:focus {
+    outline: none;
   }
+}
 </style>
