@@ -4,10 +4,10 @@
   </div>
 </template>
 <style lang="scss" scoped>
-  .th {
-    display: table-cell;
-    padding: .75rem;
-    vertical-align: top;
-    border-top: 1px solid #dee2e6;
-  }
+.th {
+  display: table-cell;
+  padding: 0.75rem;
+  vertical-align: top;
+  border-top: 1px solid #dee2e6;
+}
 </style>
