@@ -4,7 +4,7 @@
   </div>
 </template>
 <style lang="scss" scoped>
-  .thead {
-    display: table-header-group;
-  }
+.thead {
+  display: table-header-group;
+}
 </style>
