@@ -1,24 +1,27 @@
-import Vue, { CreateElement, Component, AsyncComponent } from 'vue';
-import Header from 'core/Header.vue';
-import Footer from 'core/Footer.vue';
+import { defineComponent, Component, h } from 'vue';
+import Header from '@/components/core/Header.vue';
+import Footer from '@/components/core/Footer.vue';
 
-export default function pageLayout(Component: Component | AsyncComponent) {
-  const content = (h: CreateElement) => h('div', {
-    class: 'layout__content',
-  },
-  [
-    h(Component),
-  ]);
-  return Vue.component('pageLayout', {
-    render(h) {
-      return h('main', {
-        class: 'container layout',
+export default function PageLayout(component: Component) {
+  const content = () =>
+    h(
+      'div',
+      {
+        class: 'layout__content',
       },
-      [
-        h(Header),
-        content(h),
-        h(Footer),
-      ]);
+      [h(component)],
+    );
+
+  return defineComponent({
+    name: 'PageLayout',
+    render() {
+      return h(
+        'main',
+        {
+          class: 'container layout',
+        },
+        [h(Header), content(), h(Footer)],
+      );
     },
   });
 }
