@@ -1,4 +1,4 @@
-import getLanguage from '@utils/getLanguage';
+import getLanguage from 'utils/getLanguage';
 
 // eslint-disable-next-line import/prefer-default-export
 export const LANG = getLanguage();
