@@ -1,10 +1,7 @@
 <template>
   <section class="container-fluid admin">
     <div class="admin__wrapper">
-      <Sidebar
-        :title="'Dashboard'"
-        :links="links"
-      />
+      <Sidebar :title="'Dashboard'" :links="links" />
       <div class="admin__content">
         <router-view />
       </div>
@@ -12,12 +9,10 @@
   </section>
 </template>
 
-<script lang="ts">
-import Vue from 'vue';
+<script>
+import Sidebar from 'core/Sidebar/Sidebar';
 
-import Sidebar from '@/components/core/Sidebar';
-
-export default Vue.extend({
+export default {
   components: {
     Sidebar,
   },
@@ -39,31 +34,21 @@ export default Vue.extend({
       ],
     };
   },
-});
+};
 </script>
 
 <style lang="scss">
-  @import '~styles/mixins';
-  @import '~styles/variables';
-
-  .admin {
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-
-    @include element(wrapper) {
-      flex-grow: 1;
-    }
-
-    @include element(content) {
-      padding-left: 0;
-      width: 100%;
-
-      @include media($md){
-        padding-left: 15rem;
-      }
-    }
-
+@import '~styles/mixins';
+.admin {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  @include element(wrapper) {
+    flex-grow: 1;
   }
-
+  @include element(content) {
+    padding-left: 15rem;
+    width: 100%;
+  }
+}
 </style>
