@@ -5,14 +5,12 @@
   </section>
 </template>
 
-<script lang="ts">
-import Vue from 'vue';
-
-export default Vue.extend({
+<script>
+export default {
   data() {
     return {
       title: 'Products',
     };
   },
-});
+};
 </script>
