@@ -6,12 +6,9 @@
 </template>
 
 <script>
-
 export default {
   name: 'Users',
 };
 </script>
 
-<style lang="scss">
-
-</style>
+<style lang="scss"></style>
