@@ -1,22 +1,12 @@
 <template>
   <div>
-    <nav
-      v-if="!isLoading"
-      :class="{'fadeIn': !isLoading}"
-      class="animated"
-    >
-      <Button @click="handleCreateUser">
-        Create User
-      </Button>
+    <nav v-if="!isLoading" :class="{ fadeIn: !isLoading }" class="animated">
+      <Button @click="handleCreateUser"> Create User </Button>
     </nav>
     <div class="users-table__spinner">
       <Spinner v-if="isLoading" />
     </div>
-    <div
-      v-if="!isLoading"
-      :class="{'fadeIn': !isLoading}"
-      class="animated"
-    >
+    <div v-if="!isLoading" :class="{ fadeIn: !isLoading }" class="animated">
       <Table>
         <Thead slot="thead">
           <Tr>
@@ -26,10 +16,7 @@
           </Tr>
         </Thead>
         <Tbody slot="tbody">
-          <Tr
-            v-for="user in users"
-            :key="user.id"
-          >
+          <Tr v-for="user in users" :key="user.id">
             <Th>{{ user.firstName }}</Th>
             <Td>{{ user.lastName }}</Td>
             <Td>{{ user.shop }}</Td>
@@ -40,24 +27,17 @@
   </div>
 </template>
 
-<script lang="ts">
-import Vue from 'vue';
+<script>
 import { createNamespacedHelpers } from 'vuex';
 
-import Button from '@/components/base/Button';
-import Spinner from '@/components/base/Spinner';
+import Button from 'components/base/Button';
+import Spinner from 'components/base/Spinner';
 
-import Table, {
-  Thead,
-  Tbody,
-  Tr,
-  Th,
-  Td,
-} from '@/components/base/Table';
+import Table, { Thead, Tbody, Tr, Th, Td } from 'components/base/Table';
 
 const { mapState, mapActions } = createNamespacedHelpers('users');
 
-export default Vue.extend({
+export default {
   name: 'UsersTable',
   components: {
     Table,
@@ -70,26 +50,25 @@ export default Vue.extend({
     Spinner,
   },
   computed: mapState(['users', 'isLoading']),
-  created(): void {
+  created() {
     this.fetchUsers();
   },
   methods: {
-    handleCreateUser(): void {
+    handleCreateUser() {
       this.$router.push('/admin/users/new');
     },
     ...mapActions(['fetchUsers']),
   },
-});
+};
 </script>
 
 <style lang="scss">
-  @import '~styles/functions/px-to-rem';
-  @import '~styles/mixins';
+@import '~styles/functions/px-to-rem';
+@import '~styles/mixins';
 
-  .users-table {
-    @include element(spinner) {
-      text-align: center;
-    }
+.users-table {
+  @include element(spinner) {
+    text-align: center;
   }
-
+}
 </style>
