@@ -1,7 +1,4 @@
 import * as cartActionTypes from './modules/cart/actionTypes';
 import * as productsActionTypes from './modules/products/actionTypes';
 
-export {
-  cartActionTypes,
-  productsActionTypes,
-};
+export { cartActionTypes, productsActionTypes };
