@@ -1,6 +1,6 @@
 import { IProduct } from '../products/products.types';
 
 export interface ICartState {
-  count: number,
-  products: Array<IProduct>,
+  count: number;
+  products: Array<IProduct>;
 }
