@@ -1,9 +1,6 @@
 import { MutationTree } from 'vuex';
 
-import {
-  INCREMENT_CART,
-  ADD_TO_CART,
-} from './actionTypes';
+import { INCREMENT_CART, ADD_TO_CART } from './actionTypes';
 
 import { ICartState } from './cart.types';
 
@@ -14,10 +11,7 @@ const cartMutations: MutationTree<ICartState> = {
   },
   [ADD_TO_CART](state, product) {
     // eslint-disable-next-line no-param-reassign
-    state.products = [
-      ...state.products,
-      product,
-    ];
+    state.products = [...state.products, product];
   },
 };
 
