@@ -1,10 +1,10 @@
 import { GetterTree } from 'vuex';
 
-import { IRootState } from '@/store/store.types'; 
+import { IRootState } from '../../store.types';
 import { IUsersState } from './users.type';
 
 const usersGetters: GetterTree<IUsersState, IRootState> = {
-  getUsers: state => () => state.users,
+  getUsers: (state) => () => state.users,
 };
 
 export default usersGetters;
