@@ -5,7 +5,7 @@ import getters from './getters';
 import actions from './actions';
 import mutations from './mutations';
 
-import { IRootState } from '@/store/store.types'; 
+import { IRootState } from '../../store.types';
 import { IUsersState } from './users.type';
 
 const usersModule: Module<IUsersState, IRootState> = {
