@@ -1,12 +1,12 @@
 export interface IUser {
-  firstName: string,
-  lastName: string,
-  shop: string,
+  firstName: string;
+  lastName: string;
+  shop: string;
 }
 
 export interface IUsersState {
-  user: IUser,
-  users: Array<IUser>,
-  isLoading: boolean,
-  error: string,
+  user: IUser;
+  users: Array<IUser>;
+  isLoading: boolean;
+  error: string;
 }
