@@ -1,7 +1,5 @@
-const language = (window.navigator.language);
+const language = window.navigator.language;
 
-const getLanguage = (): string => (
-  language.split('-')[0] || 'en'
-);
+const getLanguage = (): string => language.split('-')[0] || 'en';
 
 export default getLanguage;
